carbon-detection/requirements.txt
joblib==0.14.0
scikit-learn==0.21.3  
Pillow==6.2.0
numpy==1.17.2
opencv-python==4.1.1.26
